+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/me/MyBusinessUnitDraftImpl.java
@@ -30,7 +30,7 @@ public class MyBusinessUnitDraftImpl implements MyBusinessUnitDraft, ModelBase {
 
     private String contactEmail;
 
-    private com.commercetools.api.models.type.CustomFields custom;
+    private com.commercetools.api.models.type.CustomFieldsDraft custom;
 
     private java.util.List<com.commercetools.api.models.common.BaseAddress> addresses;
 
@@ -49,7 +49,7 @@ public class MyBusinessUnitDraftImpl implements MyBusinessUnitDraft, ModelBase {
     MyBusinessUnitDraftImpl(@JsonProperty("key") final String key,
             @JsonProperty("unitType") final com.commercetools.api.models.business_unit.BusinessUnitType unitType,
             @JsonProperty("name") final String name, @JsonProperty("contactEmail") final String contactEmail,
-            @JsonProperty("custom") final com.commercetools.api.models.type.CustomFields custom,
+            @JsonProperty("custom") final com.commercetools.api.models.type.CustomFieldsDraft custom,
             @JsonProperty("addresses") final java.util.List<com.commercetools.api.models.common.BaseAddress> addresses,
             @JsonProperty("shippingAddresses") final java.util.List<Integer> shippingAddresses,
             @JsonProperty("defaultShippingAddress") final Integer defaultShippingAddress,
@@ -109,7 +109,7 @@ public class MyBusinessUnitDraftImpl implements MyBusinessUnitDraft, ModelBase {
      *  <p>Custom Fields for the Business Unit.</p>
      */
 
-    public com.commercetools.api.models.type.CustomFields getCustom() {
+    public com.commercetools.api.models.type.CustomFieldsDraft getCustom() {
         return this.custom;
     }
 
@@ -165,7 +165,7 @@ public class MyBusinessUnitDraftImpl implements MyBusinessUnitDraft, ModelBase {
         this.contactEmail = contactEmail;
     }
 
-    public void setCustom(final com.commercetools.api.models.type.CustomFields custom) {
+    public void setCustom(final com.commercetools.api.models.type.CustomFieldsDraft custom) {
         this.custom = custom;
     }